Fixes #3770
