+++ radio/src/myeeprom.h
@@ -47,12 +47,6 @@
   #define EEPROM_VER             216
 #endif
 
-#if defined(PCBSTD)
-  #define GVAR_VALUE(x, p)             g_model.gvars[x]
-#else
-  #define GVAR_VALUE(gv, fm)           g_model.flightModeData[fm].gvars[gv]
-#endif
-
 #define GET_PPM_POLARITY(idx)             g_model.moduleData[idx].ppm.pulsePol
 #define GET_PPM_DELAY(idx)                (g_model.moduleData[idx].ppm.delay * 50 + 300)
 #define SET_DEFAULT_PPM_FRAME_LENGTH(idx) g_model.moduleData[idx].ppm.frameLength = 4 * max((int8_t)0, g_model.moduleData[idx].channelsCount)
